Implement rule evaluation and scoring (#11)
Stitch together the manifest (collection of models) and the rule
registry (collection of rules) to evaluate every model with every rule
defined.

For every model, its score is computed. The aggregation of these scores
is also computed in the project score.\

Add a human-readable formatter to display those results upon CLI
invocation.