Implement --select option (#34)
When running `dbt-score lint`, the `-s/--select` option allows to only
lint a selection of models.

When the selection is "simple", i.e. one or multiple model names, a
simple filtering of known models is applied. This allows to run it fast.

When the selection is more complex, using dbt's node selection syntax
(https://docs.getdbt.com/reference/node-selection/syntax), the filtering
is deferred to dbt (`dbt ls`).