Add null check before calling project_evaluated (#64)
This PR adds a null check in the evaluate method of the Evaluation class
before calling project_evaluated on the formatter. This ensures that
project_evaluated is only called if there are models to evaluate, thus
skipping the output when no model is found.

#### Changes Made
Added a null check in the evaluate method of the Evaluation class.