Only output failing models and violated rules per default in HumanReadableFormatter (#77)
Fixes #71 

It changes the default behavior of the HumanReadableFormatter:
When running `dbt-score lint` it will now only show failing models
and/or rules which were violated.
To have all output printed (exactly the way it was before this PR) the
user has now to pass the `--show_all` flag to `dbt-score lint`.
It also considers the `fail_any_model_under` option. 

Furthermore, I had to adjust some of tests since the default behavior of
the HumanReadableFormatter is now changed - let me know if I would need
to expand this.

The docs also reflect this change by slightly tweaking the first example
in `index.md`

In this current form, both rule violations and failing models are
filtered through the same argument. This is based on @matthieucan. 's
comment on
https://github.com/PicnicSupermarket/dbt-score/issues/71#issuecomment-2383106560

---------

Co-authored-by: Jochem van Dooren <jochemmathijs.vandooren@teampicnic.com>
Co-authored-by: Matthieu Caneill <matthieucan@users.noreply.github.com>