Adding children in addition to parents (#113)
As discussed
[here](https://github.com/PicnicSupermarket/dbt-score/issues/111),
adding `children` to models / sources / snapshots. This is a simple
addition to the logic that populates parents; basically, if `x` is one
of `y`'s parents, then `y` is one of `x`'s children.

This allows writing rules like:
```
@rule(severity=Severity.CRITICAL)
def children_of_beta_models_should_be_beta_models(
    model: Model,
) -> RuleViolation | None:
    """Children of beta models should be beta models"""
    if "beta" in model.tags:
        for child in model.children:
            if "beta" not in child.tags:
                return RuleViolation(
                    message=f"Beta model {model.name} has a child that isn't a beta model {child.name}"
                )
    return None
```

This will need to be updated slightly for new evaluable types, but that
should be easy.

---------

Co-authored-by: Jochem van Dooren <jochemmathijs.vandooren@teampicnic.com>