Display dbt parse errors (#126)
When running dbt-score, if dbt parse fails, the error message does not
reflect the root cause of the failure. This PR attaches the root cause
to the DbtParseException object and prints it to the console.

Here’s an example of how it looks:

```shell
$ dbt-score lint -p

ERROR:dbt_score.cli:dbt failed to parse project. dbt parse failed.

Root cause: Compilation Error
  Model 'model.edge.example_incremental_model' (models/example_incremental_model/example_incremental_model.sql) depends on a node named 'example_overwrite_not_exists_model' which was not found
```